NutriCare_Project/api/requirements.txt
flask 
Flask-JWT-Extended
mysql-connector-python
flask_bcrypt